CM - Endpoint de lsitagem de alunos para acompanhamento de frequência.

+++ src/SME.SGP.Infra/Dtos/AcompanhamentoFrequencia/ObterFrequenciaAlunosPorBimestreDto.cs
@@ -4,6 +4,6 @@
     {
         public long ComponenteCurricularId { get; set; }
         public long TurmaId { get; set; }
-        public short? Bimestre { get; set; }
+        public int? Bimestre { get; set; }
     }
 }
